move get-recipe to get-history
