+++ lib/get-history.js
@@ -1,26 +1,26 @@
-const {markdown} = require("markdown");
-const fs = require("fs");
-const path = require("path");
+const markdown = require('markdown-it')('commonmark');
+const fs = require('fs');
+const path = require('path');
 
 function getFileNames(pathName) {
   return fs.readdirSync(path.resolve(__dirname, pathName))
     .map(filename => {
-        return path.join(__dirname, pathName, filename);
+      return path.join(__dirname, pathName, filename);
     });
 }
 
-const allPeople = getFileNames("../history/people");
-const allEvents = getFileNames("../history/events");
-const allTrends = getFileNames("../history/trends");
+const allPeople = getFileNames('../history/people');
+const allEvents = getFileNames('../history/events');
+const allTrends = getFileNames('../history/trends');
 const allHistory = allPeople.concat(allEvents).concat(allTrends);
 console.log(allHistory);
 let lastFile;
 
 function readAndParse(file) {
   return new Promise((resolve, reject) => {
-    fs.readFile(file, {encoding: "utf-8"}, (err, text) => {
+    fs.readFile(file, {encoding: 'utf-8'}, (err, text) => {
       if (err) return reject(err);
-      resolve(markdown.toHTML(text));
+      resolve(markdown.render(text));
     });
   });
 }
@@ -30,7 +30,7 @@ module.exports = function getRandomRecipe() {
     let files;
     if (lastFile) {
       files = [].concat(allHistory);
-      files.splice(allHistory.indexOf(lastFile), 1)
+      files.splice(allHistory.indexOf(lastFile), 1);
     } else {
       files = allHistory;
     }
